feat: 헤더 컴포넌트 분리 및 아이콘 변경
