+++ src/app/components/Navi.tsx
@@ -2,13 +2,8 @@
 
 import Link from 'next/link'
 import NavList from './NavList'
-import { useAuth } from '@/hooks/useAuth'
+import Header from './Header'
 
-interface NaviProps {
-  name: string
-  href: string
-  icon: string
-}
 
 const mainMenu = [
   { name: '필기구', href: '/products/writing' },
@@ -17,18 +12,7 @@ const mainMenu = [
   { name: '소품/액세서리', href: '/products/accessory' },
 ]
 
-export default function Header() {
-  const { isLogin, handleLogout } = useAuth()
-
-  const convenienceMenu = [
-    { name: '검색', onClick: () => console.log('검색 클릭'), icon: '🔍' },
-    isLogin && { name: '마이페이지', href: '/mypage/consumer', icon: '👤' },
-    { name: '장바구니', href: '/cart', icon: '🛒' },
-    isLogin
-      ? { name: '로그아웃', onClick: handleLogout, icon: '🔓' }
-      : { name: '로그인', href: '/login', icon: '🔒' },
-  ].filter((item): item is NaviProps => item !== false)
-
+export default function Navi() {
   return (
     <>
       {/* 쿠폰 */}
@@ -66,8 +50,8 @@ export default function Header() {
             mainMenu
           />
 
-          {/* 유저 메뉴 */}
-          <NavList label="convenience-menu" items={convenienceMenu} />
+          {/* 헤더 */}
+          <Header />
         </div>
       </header>
     </>
